requirements.txt
huggingface_hub==0.23.4
diffusers==0.28.2
transformers==4.40.2
accelerate==0.30.1
peft==0.11.1
safetensors==0.4.3
flask==3.0.3
Pillow==10.3.0
